// src/common/enum/category.enum.ts
export enum Category {
  FICTION = 'Fiction',
  NON_FICTION = 'Non-Fiction',
  SCIENCE = 'Science',
  HISTORY = 'History',
  BIOGRAPHY = 'Biography',
  FANTASY = 'Fantasy',
  MYSTERY = 'Mystery',
  ROMANCE = 'Romance',
  HORROR = 'Horror',
  SELF_HELP = 'Self-Help',
  HEALTH = 'Health',
  TRAVEL = 'Travel',
  CHILDREN = 'Children',
  YOUNG_ADULT = 'Young Adult',
  EDUCATION = 'Education',
}